Reorganized Files
- Moved MIDI Receiving code to its own file.
- Created an empty MIDI Creator file
- Created separate folder for utility code that should not be tied to a single page.
